backend/requirements.txt
asgiref==3.10.0
distlib==0.4.0
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
filelock==3.20.0
mysql-connector-python==9.4.0
mysqlclient==2.2.7
numpy==2.3.5
pandas==2.3.3
platformdirs==4.5.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
virtualenv==20.35.4
